+++ src/collective/eeafaceted/batchactions/browser/static/batch_actions.js
@@ -28,7 +28,7 @@ collective_batch_actions.init_button = function () {
     uids_input.val(uids);
     ba_form.action = document.batch_actions[form_id] + '?referer=' + referer;
     if ($(ba_form).hasClass('do-overlay')) {
-      collective_batch_actions.initializeOverlays('#'+form_id);
+      collective_batch_actions.initializeOverlays(ba_form);
     }
     else {
       ba_form.submit();
@@ -37,9 +37,9 @@ collective_batch_actions.init_button = function () {
   });
 };
 
-collective_batch_actions.initializeOverlays = function (form_id) {
+collective_batch_actions.initializeOverlays = function (ba_form) {
     // Add batch actions popup
-    $(form_id).prepOverlay({
+    $(ba_form).prepOverlay({
         api: true,
         subtype: 'ajax',
         closeselector: '[name="form.buttons.cancel"]',